working and refactoring
