+++ src/components/parrotList.js
@@ -6,15 +6,18 @@ import styles from '../../styles';
 const ParrotList = ({ navigation }) => {
 	const [parrots, setParrots] = useState([]);
 
-	useEffect(async () => {
-		console.log('fetch data in use effect');
-		const res = await fetch(`http://localhost:3000/api/parrots`, {
-			method: 'GET',
-		})
-			.then((response) => response.json())
-			.then((data) => {
-				setParrots(data);
-			});
+	useEffect(() => {
+		const fetchParrots = async () => {
+			console.log('fetch data in use effect');
+			const res = await fetch(`http://localhost:3000/api/parrots`, {
+				method: 'GET',
+			})
+				.then((response) => response.json())
+				.then((data) => {
+					setParrots(data);
+				});
+		};
+		fetchParrots();
 	}, []);
 
 	return (
@@ -24,6 +27,7 @@ const ParrotList = ({ navigation }) => {
 				renderItem={({ item }) => (
 					<Parrot
 						key={item._id}
+						id={item._id}
 						name={item.name}
 						imgUrl={item.imageUrl ? item.imageUrl : 'https://picsum.photos/200'}
 						navigation={navigation}
